requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-ollama
langchain-chroma
chromadb
sentence-transformers
pandas
python-dotenv
typer
rich
requests
streamlit
ragas
openai
rank-bm25
